:memo: Update README & rename few files
