+++ src/core/helpers/general.py
